# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * VERIF-AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserSyncInput {
  clerkId: string;
  email: string;
  role: string;
  displayName: string;
  /** @nullable */
  companyName?: string | null;
}

export interface UserSyncResult {
  userId: number;
  role: string;
}

export interface UserProfile {
  id: number;
  clerkId: string;
  email: string;
  role: string;
  displayName: string;
  /** @nullable */
  companyName?: string | null;
  /** @nullable */
  college?: string | null;
  skills?: string[];
  /** @nullable */
  linkedinUrl?: string | null;
  /** @nullable */
  githubUrl?: string | null;
  /** @nullable */
  portfolioUrl?: string | null;
  /** @nullable */
  resumeUrl?: string | null;
  /** @nullable */
  resumeFilename?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  appliedJobIds?: number[];
  createdAt: string;
}

export interface Job {
  id: number;
  title: string;
  description: string;
  company: string;
  location: string;
  jobType: string;
  /** @nullable */
  salary?: string | null;
  skills: string[];
  recruiterId: number;
  postedAt: string;
  isActive: boolean;
  applicantCount: number;
}

export interface JobInput {
  title: string;
  description: string;
  company: string;
  location: string;
  jobType: string;
  /** @nullable */
  salary?: string | null;
  skills: string[];
}

export interface JobFilters {
  skills: string[];
  jobTypes: string[];
}

export interface Application {
  id: number;
  jobId: number;
  studentId: number;
  status: string;
  appliedAt: string;
  job?: Job;
}

export interface Certificate {
  id: number;
  filename: string;
  url: string;
  uploadedAt: string;
}

export interface StudentProfile {
  id: number;
  displayName: string;
  email: string;
  role: string;
  /** @nullable */
  college?: string | null;
  skills?: string[];
  /** @nullable */
  linkedinUrl?: string | null;
  /** @nullable */
  githubUrl?: string | null;
  /** @nullable */
  portfolioUrl?: string | null;
  /** @nullable */
  resumeUrl?: string | null;
  /** @nullable */
  resumeFilename?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  appliedJobIds?: number[];
  certificates?: Certificate[];
  createdAt: string;
}

export interface ApplicationWithStudent {
  id: number;
  jobId: number;
  studentId: number;
  status: string;
  appliedAt: string;
  student?: StudentProfile;
  job?: Job;
}

export interface ApplicationStatusInput {
  status: string;
}

export interface StudentProfileInput {
  displayName?: string;
  /** @nullable */
  college?: string | null;
  skills?: string[];
  /** @nullable */
  linkedinUrl?: string | null;
  /** @nullable */
  githubUrl?: string | null;
  /** @nullable */
  portfolioUrl?: string | null;
}

export interface StudentStats {
  totalApplications: number;
  pendingApplications: number;
  shortlistedApplications: number;
  profileCompletion: number;
  recentApplications?: Application[];
}

export interface RecruiterProfile {
  id: number;
  displayName: string;
  email: string;
  role: string;
  /** @nullable */
  companyName?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  createdAt: string;
}

export interface RecruiterProfileInput {
  displayName?: string;
  /** @nullable */
  companyName?: string | null;
}

export interface RecruiterStats {
  totalJobs: number;
  totalApplications: number;
  pendingApplications: number;
  shortlistedApplications: number;
  recentApplications?: ApplicationWithStudent[];
}

export interface AICheckResult {
  status: string;
  detail: string;
}

export interface AIVerificationResult {
  id: number;
  studentId: number;
  trustScore: number;
  verdict: string;
  summary: string;
  strengths?: string[];
  redFlags?: string[];
  educationCheck?: AICheckResult;
  experienceCheck?: AICheckResult;
  githubCheck?: AICheckResult;
  certificatesCheck?: AICheckResult;
  recommendation?: string;
  analyzedAt: string;
}

export type ListJobsParams = {
search?: string;
skills?: string;
jobType?: string;
};