# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VERIF-AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/sync:
    post:
      operationId: syncUser
      tags: [auth]
      summary: Sync authenticated user into DB
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSyncInput"
      responses:
        "200":
          description: Synced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSyncResult"
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List all active jobs
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: skills
          in: query
          schema: { type: string }
        - name: jobType
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a new job (recruiter only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
...
  /jobs/{jobId}/apply:
    post:
      operationId: applyToJob
      tags: [jobs]
      summary: Apply to a job (student only)
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Applied successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"

  # ── STUDENT ───────────────────────────────────────────────────────────────
...
  /ai/analyze/{studentId}:
    post:
      operationId: analyzeStudent
      tags: [ai]
      summary: Trigger AI verification for a student
      parameters:
        - name: studentId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: AI verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIVerificationResult"
  /ai/result/{studentId}:
    get:
      operationId: getAIResult
      tags: [ai]
      summary: Get latest AI verification result for a student
      parameters:
        - name: studentId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: AI result or null
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIVerificationResult"
components:
  schemas:
...
    UserSyncInput:
      type: object
      required: [clerkId, email, role, displayName]
      properties:
        clerkId: { type: string }
        email: { type: string }
        role: { type: string }
        displayName: { type: string }
        companyName: { type: ["string", "null"] }
    UserSyncResult:
      type: object
      required: [userId, role]
      properties:
        userId: { type: integer }
        role: { type: string }
...
    Job:
      type: object
      required: [id, title, description, company, location, jobType, skills, recruiterId, postedAt, isActive, applicantCount]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        company: { type: string }
        location: { type: string }
        jobType: { type: string }
        salary: { type: ["string", "null"] }
        skills: { type: array, items: { type: string } }
        recruiterId: { type: integer }
        postedAt: { type: string }
        isActive: { type: boolean }
        applicantCount: { type: integer }
    JobInput:
      type: object
      required: [title, description, company, location, jobType, skills]
      properties:
        title: { type: string }
        description: { type: string }
        company: { type: string }
        location: { type: string }
        jobType: { type: string }
        salary: { type: ["string", "null"] }
        skills: { type: array, items: { type: string } }
...
    Application:
      type: object
      required: [id, jobId, studentId, status, appliedAt]
      properties:
        id: { type: integer }
        jobId: { type: integer }
        studentId: { type: integer }
        status: { type: string }
        appliedAt: { type: string }
        job:
          $ref: "#/components/schemas/Job"
...
    AICheckResult:
      type: object
      required: [status, detail]
      properties:
        status: { type: string }
        detail: { type: string }
    AIVerificationResult:
      type: object
      required: [id, studentId, trustScore, verdict, summary, analyzedAt]
      properties:
        id: { type: integer }
        studentId: { type: integer }
        trustScore: { type: integer }
        verdict: { type: string }
        summary: { type: string }
        strengths: { type: array, items: { type: string } }
        redFlags: { type: array, items: { type: string } }
        educationCheck:
          $ref: "#/components/schemas/AICheckResult"
        experienceCheck:
          $ref: "#/components/schemas/AICheckResult"
        githubCheck:
          $ref: "#/components/schemas/AICheckResult"
        certificatesCheck:
          $ref: "#/components/schemas/AICheckResult"
        recommendation: { type: string }
        analyzedAt: { type: string }